refactor(test): Move helper functions to separate file
